Relabel app to grommunio-chat
Change text of about dialog

Change bundle identifier to grommunio

Update ios icons

Change ios icons to white background and remove redundant ones

Exchange logos with grammm logos

Relabel about dialog

Change system message profile picture to grammm logo

Update ToS and privacy links

Remove dash from ios display name

Remove dash from android app name

Change bundle identifier

Exchange logos with transparent ones

Exchange icon with more centralized version

Change example URL

Change example urls to https://example.com/chat

Remove grammm

Rename package to chat.grommunio.com

Sync ios folder from g-c-master branch

Resolve all rebase errors

Upgrade to targetSdkVersion 30

Exchange new system message component

Rename grammm to grommunio

Exchange 'Mattermost' in translations

+++ ios/Mattermost/UIImage+ImageEffects.m
@@ -127,7 +127,7 @@
 {
   const CGFloat EffectColorAlpha = 0.6;
   UIColor *effectColor = tintColor;
-  size_t componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
+  int componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
   if (componentCount == 2) {
     CGFloat b;
     if ([tintColor getWhite:&b alpha:NULL]) {
